requirements.txt
Flask>=2.0
requests>=2.25
python-dotenv>=0.21.0
gunicorn>=20.1.0
